Rename files.
Reviewed by @tolmasky.
