+++ packages/babel-parser/src/parse-error/module-errors.js
