+++ packages/babel-parser/src/parse-error/strict-mode-errors.js
